feat: first release version determined by the bump, not hard-coded 0.1.0 (#3)
## Related GitHub Issues
<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

I have a project that I'm working on, and I want the very first release
to be 1.0.0, but with this script, it doesn't allow it at the moment.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

The first release version is determined by the version bump, not a
hard-coded 0.1.0.

## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [X] Added automated tests. 
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->